Refactor emitters, extracting reusable rewriters to traits
