+++ src/main/php/lang/ast/emit/HHVM320.class.php
@@ -7,6 +7,9 @@ use lang\ast\Emitter;
  * types.
  */
 class HHVM320 extends Emitter {
+  use OmitPropertyTypes, OmitConstModifiers;
+  use RewriteNullCoalesceAssignment, RewriteLambdaExpressions, RewriteMultiCatch;
+
   protected $unsupported= [
     'object'   => 72,
     'void'     => 71,
@@ -17,38 +20,9 @@ class HHVM320 extends Emitter {
   protected function emitParameter($parameter) {
     if ($parameter->variadic) {
       $this->out->write('... $'.$parameter->name);
+      $this->locals[$parameter->name]= true;
     } else {
-      if ($parameter->type && $t= $this->paramType($parameter->type->literal())) {
-        $this->out->write($t.' ');
-      }
-      $this->out->write(($parameter->reference ? '&' : '').'$'.$parameter->name);
+      parent::emitParameter($parameter);
     }
-    if ($parameter->default) {
-      $this->out->write('=');
-      $this->emit($parameter->default);
-    }
-    $this->locals[$parameter->name]= true;
-  }
-
-  protected function emitCatch($catch) {
-    if (empty($catch->types)) {
-      $this->out->write('catch(\\Throwable $'.$catch->variable.') {');
-    } else {
-      $last= array_pop($catch->types);
-      $label= sprintf('c%u', crc32($last));
-      foreach ($catch->types as $type) {
-        $this->out->write('catch('.$type.' $'.$catch->variable.') { goto '.$label.'; }');
-      }
-      $this->out->write('catch('.$last.' $'.$catch->variable.') { '.$label.':');
-    }
-
-    $this->emit($catch->body);
-    $this->out->write('}');
-  }
-
-  protected function emitConst($const) {
-    $this->out->write('const '.$const->name.'=');
-    $this->emit($const->expression);
-    $this->out->write(';');
   }
 }
