+++ src/main/php/lang/ast/emit/PHP70.class.php
@@ -1,44 +1,20 @@
 <?php namespace lang\ast\emit;
 
+use lang\ast\Emitter;
+
 /**
  * PHP 7.0 syntax
  *
- * @see  https://wiki.php.net/rfc/generator-delegation - Not yet implemented
- * @see  https://wiki.php.net/rfc/generator-return-expressions - Not yet implemented
- * @see  https://wiki.php.net/rfc/anonymous_classes
- * @see  https://wiki.php.net/rfc/return_types
- * @see  https://wiki.php.net/rfc/isset_ternary
- * @see  https://wiki.php.net/rfc/uniform_variable_syntax
- * @see  https://wiki.php.net/rfc/group_use_declarations
- * @see  https://wiki.php.net/rfc/scalar_type_hints_v5
+ * @see  https://wiki.php.net/rfc#php_70
  */
-class PHP70 extends \lang\ast\Emitter {
+class PHP70 extends Emitter {
+  use OmitPropertyTypes, OmitConstModifiers;
+  use RewriteNullCoalesceAssignment, RewriteLambdaExpressions, RewriteMultiCatch;
+
   protected $unsupported= [
     'object'   => 72,
     'void'     => 71,
     'iterable' => 71,
     'mixed'    => null,
   ];
-
-  protected function emitCatch($catch) {
-    if (empty($catch->types)) {
-      $this->out->write('catch(\\Throwable $'.$catch->variable.') {');
-    } else {
-      $last= array_pop($catch->types);
-      $label= sprintf('c%u', crc32($last));
-      foreach ($catch->types as $type) {
-        $this->out->write('catch('.$type.' $'.$catch->variable.') { goto '.$label.'; }');
-      }
-      $this->out->write('catch('.$last.' $'.$catch->variable.') { '.$label.':');
-    }
-
-    $this->emit($catch->body);
-    $this->out->write('}');
-  }
-
-  protected function emitConst($const) {
-    $this->out->write('const '.$const->name.'=');
-    $this->emit($const->expression);
-    $this->out->write(';');
-  }
 }
