+++ src/main/php/lang/ast/emit/PHP71.class.php
@@ -1,16 +1,16 @@
 <?php namespace lang\ast\emit;
 
+use lang\ast\Emitter;
+
 /**
  * PHP 7.1 syntax
  *
- * @see  https://wiki.php.net/rfc/nullable_types - Not yet implemented!
- * @see  https://wiki.php.net/rfc/short_list_syntax - Not yet implemented!
- * @see  https://wiki.php.net/rfc/class_const_visibility - Not yet implemented!
- * @see  https://wiki.php.net/rfc/multiple-catch
- * @see  https://wiki.php.net/rfc/void_return_type
- * @see  https://wiki.php.net/rfc/iterable
+ * @see  https://wiki.php.net/rfc#php_71
  */
-class PHP71 extends \lang\ast\Emitter {
+class PHP71 extends Emitter {
+  use OmitPropertyTypes;
+  use RewriteNullCoalesceAssignment, RewriteLambdaExpressions;
+
   protected $unsupported= [
     'object'   => 72,
     'mixed'    => null,
