+++ src/main/php/lang/ast/emit/PHP72.class.php
@@ -1,11 +1,16 @@
 <?php namespace lang\ast\emit;
 
+use lang\ast\Emitter;
+
 /**
  * PHP 7.2 syntax
  *
- * @see  https://wiki.php.net/rfc/object-typehint
+ * @see  https://wiki.php.net/rfc#php_72
  */
-class PHP72 extends \lang\ast\Emitter {
+class PHP72 extends Emitter {
+  use OmitPropertyTypes;
+  use RewriteNullCoalesceAssignment, RewriteLambdaExpressions;
+
   protected $unsupported= [
     'mixed'    => null,
   ];
